requirements.txt
aiogram==3.3.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
SQLAlchemy==2.0.27
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0